origin/ReplaceAlamofire: move "Map" methods to test (test class are the only ones used those methods)
